requirements.txt
# GAIA Framework - Dependencies
# Python 3.10+

# Core dependencies
click>=8.1.0
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.28.0

# CLI
rich>=13.0.0
questionary>=2.0.0

# Knowledge
networkx>=3.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
